requirements/common.txt
Django==1.10.5
djangorestframework==3.5.4
django-mptt==0.8.7
django-filter==1.0.1
django-js-reverse==0.7.3
Pillow==4.0.0